+++ source/service/builder/lib/make-config.js
@@ -36,7 +36,7 @@ export default function(config, sections, assets)
     lines.push(`export let sections = ${jss(sections)}`);
     lines.push(`export let assets = ${jss(assets)}`);
 
-    lines.push(`export let cobe = await mapExtensions(${jss(cobe)})`);
+    lines.push(`export let cobe = await mapExtensions(${jss(cobe, [ 're' ])})`);
 
     lines.push(`let mapFn = async k => sections[k].cobe &&= await mapExtensions(sections[k].cobe)`);
     lines.push(`await Promise.all(Object.keys(sections).map(mapFn));`);